move get query hash function to util
